# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API en la cual podemos consultar cotizaciones de diferentes monedas",
        "title": "Criptomoneda API",
        "contact": {}
    },
    "paths": {
        "/cotizacion": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cotizacion"
                ],
                "summary": "Usuario hace cotizacion de moneda manualmente",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Usuario que cotizara",
                        "name": "id-usuario",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Simbolo de la moneda que cotizara",
                        "name": "simbolo",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Valor que cotizara",
                        "name": "precio",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Fecha de la cotizacion",
                        "name": "fecha",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/cotizacion/{id}": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cotizacion"
                ],
                "summary": "Usuario cambia cotizacion de moneda manualmente",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id de cotizacion",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Usuario que hace los cambios",
                        "name": "id-usuario",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Cotizacion Actualizada",
                        "name": "cambios",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ports.CotizacionPut"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cotizacion"
                ],
                "summary": "Usuario elimina cotizacion de moneda manualmente",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Usuario que elimina",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "cotizacion a eliminar",
                        "name": "id-cotizacion",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/cotizaciones": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cotizacion"
                ],
                "summary": "Retorna las cotizaciones paginadas segun los filtros",
                "parameters": [
                    {
                        "type": "string",
                        "description": "simbolos de las monedas que queremos separados por espacios",
                        "name": "monedas",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "date-time",
                        "description": "Fecha desde la cual quiero obtener cotizaciones (YYYY-MM-DD HH:MM:SS)",
                        "name": "fecha_inicial",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "date-time",
                        "description": "Fecha hasta la cual quiero obtener cotizaciones (YYYY-MM-DD HH:MM:SS)",
                        "name": "fecha_final",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "El tamaño de las paginas, como maximo es 100, el default es 50",
                        "name": "tam_paginas",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Pagina a partir de la cual sera retornado el query",
                        "name": "pagina_inicial",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "La cantidad de paginas, como maximo es 10, el default es 10",
                        "name": "cant_paginas",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Ordenar segun alguno de estos valores: fecha(default), valor, nombre",
                        "name": "orden",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Indica si es ascendente o descendente, el default es desdencente",
                        "name": "orden_direccion",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Usuario elegido",
                        "name": "usuario",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Para incluir resumen indicar el valor debe ser si",
                        "name": "resumen",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cotizacion"
                ],
                "summary": "Llama para que se haga la cotizacion de las monedas",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token de autorización",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "API que se utilizara para cotizar",
                        "name": "api",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "ports.CotizacionPut": {
            "type": "object",
            "properties": {
                "fecha": {
                    "type": "string"
                },
                "simbolo": {
                    "type": "string"
                },
                "valor": {
                    "type": "number"
                }
            }
        },
...
    }
}